+++ examples/dagster_examples_tests/stocks_tests/test_stocks_example.py
@@ -5,4 +5,4 @@ def test_basic():
     repo = partitioning_tutorial
     assert len(repo.get_all_pipelines()) == 1
     assert len(repo.schedule_defs) == 1
-    assert len(repo.partition_set_defs) == 1
+    assert len(repo.partition_set_defs) == 2  # 1 from schedule, 1 on repo