+++ python_modules/dagster/dagster/core/host_representation/__init__.py
@@ -6,12 +6,20 @@ in user processes and containers.  e.g. ExternalPipeline
 It also contains classes that represent historical representations
 that have been persisted. e.g. HistoricalPipeline
 '''
-from .external import ExternalExecutionPlan, ExternalPipeline, ExternalRepository
+from .external import (
+    ExternalExecutionPlan,
+    ExternalPartitionSet,
+    ExternalPipeline,
+    ExternalRepository,
+    ExternalSchedule,
+)
 from .external_data import (
+    ExternalPartitionSetData,
     ExternalPipelineData,
     ExternalPipelineSubsetResult,
     ExternalPresetData,
     ExternalRepositoryData,
+    ExternalScheduleData,
     external_pipeline_data_from_def,
     external_repository_data_from_def,
 )