+++ python_modules/dagster/dagster_tests/core_tests/snap_tests/snapshots/snap_test_active_data.py
@@ -8,6 +8,48 @@ snapshots = Snapshot()
 
 snapshots['test_external_repository_data 1'] = '''{
   "__class__": "ExternalRepositoryData",
+  "external_partition_set_datas": [
+    {
+      "__class__": "ExternalPartitionSetData",
+      "mode": "default",
+      "name": "a_schedule_partitions",
+      "partition_names": [
+        "2019-01-01",
+        "2019-01-02",
+        "2019-01-03",
+        "2019-01-04",
+        "2019-01-05",
+        "2019-01-06",
+        "2019-01-07",
+        "2019-01-08",
+        "2019-01-09",
+        "2019-01-10",
+        "2019-01-11",
+        "2019-01-12",
+        "2019-01-13",
+        "2019-01-14",
+        "2019-01-15",
+        "2019-01-16",
+        "2019-01-17",
+        "2019-01-18",
+        "2019-01-19",
+        "2019-01-20",
+        "2019-01-21",
+        "2019-01-22",
+        "2019-01-23",
+        "2019-01-24",
+        "2019-01-25",
+        "2019-01-26",
+        "2019-01-27",
+        "2019-01-28",
+        "2019-01-29",
+        "2019-01-30",
+        "2019-01-31"
+      ],
+      "pipeline_name": "a_pipeline",
+      "solid_subset": null
+    }
+  ],
   "external_pipeline_datas": [
     {
       "__class__": "ExternalPipelineData",
@@ -1101,6 +1143,17 @@ snapshots['test_external_repository_data 1'] = '''{
         },
         "tags": {}
       }
+    }
+  ],
+  "external_schedule_datas": [
+    {
+      "__class__": "ExternalScheduleData",
+      "cron_schedule": "0 0 * * *",
+      "mode": "default",
+      "name": "a_schedule",
+      "partition_set_name": "a_schedule_partitions",
+      "pipeline_name": "a_pipeline",
+      "solid_subset": null
     }
   ],
   "name": "repo"