+++ python_modules/dagster/dagster_tests/core_tests/snap_tests/test_active_data.py
@@ -1,4 +1,6 @@
-from dagster import ModeDefinition, PresetDefinition, pipeline, repository, solid
+from datetime import datetime
+
+from dagster import ModeDefinition, PresetDefinition, daily_schedule, pipeline, repository, solid
 from dagster.core.host_representation import (
     external_pipeline_data_from_def,
     external_repository_data_from_def,
@@ -27,12 +29,25 @@ def a_pipeline():
     a_solid()
 
 
+@daily_schedule(
+    pipeline_name='a_pipeline',
+    start_date=datetime(year=2019, month=1, day=1),
+    end_date=datetime(year=2019, month=2, day=1),
+)
+def a_schedule():
+    return {}
+
+
 def test_external_repository_data(snapshot):
     @repository
     def repo():
-        return [a_pipeline]
+        return [a_pipeline, a_schedule]
 
-    snapshot.assert_match(serialize_pp(external_repository_data_from_def(repo)))
+    external_repo_data = external_repository_data_from_def(repo)
+    assert external_repo_data.get_external_pipeline_data('a_pipeline')
+    assert external_repo_data.get_external_schedule_data('a_schedule')
+    assert external_repo_data.get_external_partition_set_data('a_schedule_partitions')
+    snapshot.assert_match(serialize_pp(external_repo_data))
 
 
 def test_external_pipeline_data(snapshot):